requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
fastapi>=0.75.2
uvicorn>=0.17.6
transformers>=4.30.0
qdrant-client>=1.12.0
python-dotenv>=0.21.1
sentence-transformers>=2.2.2
tqdm>=4.64.1
numpy>=1.20.0
pyperclip>=1.8.2
huggingface-hub>=0.16.0